% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timing.R
\name{add.timing}
\alias{add.timing}
\title{Add timing to cache}
\usage{
add.timing(x, .timing.name = getOption("dwtools.timing.name"),
  .timing.conn.name = getOption("dwtools.timing.conn.name"))
}
\description{
Add timing to cache
}
\keyword{internal}

