% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anonymize}
\alias{anonymize}
\title{Anonymize}
\usage{
anonymize(x, algo)
}
\arguments{
\item{x}{vector to anonymize}

\item{algo}{character scalar to be used for \emph{digest} function in \emph{digest} package, read \code{?digest::digest}.}
}
\value{
character vector of hashes, same length as length of \emph{x} argument.
}
\description{
Mask data using hasing function
}
\examples{
suppressPackageStartupMessages(library(dwtools))

X = dw.populate(N=1e5, scenario="star")
DT <- joinbyv(X$SALES, list(X$CUSTOMER))
print(DT)
cols_to_mask <- c("cust_code", "cust_name", "cust_mail")
DT[, (cols_to_mask) := lapply(.SD, anonymize, "crc32"), .SDcols=c(cols_to_mask)]
print(DT)
}

