% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.xts.data.table}
\alias{as.xts.data.table}
\title{as.xts.data.table}
\usage{
as.xts.data.table(x)
}
\arguments{
\item{x}{data.table to convert to xts, must have \emph{POSIXct} or \emph{Date} in the first column. All others non-numeric columns will be omitted with warning.}
}
\description{
convert data.table to xts
}
\examples{
suppressPackageStartupMessages(library(dwtools))

sample.dt <- data.table(date = as.Date((Sys.Date()-999):Sys.Date(),origin="1970-01-01"),
                        quantity = sample(10:50,1000,TRUE),
                        value = sample(100:1000,1000,TRUE))
# print dt
print(sample.dt)
# print head of xts
print(head(as.xts.data.table(sample.dt))) # xts non attached on search path
}
\seealso{
\link{as.data.table.xts}
}

