% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{auto.table.name}
\alias{auto.table.name}
\title{Auto table name generate}
\usage{
auto.table.name(x, ncol = getOption("dwtools.db.auto.table.name.ncol"),
  nchar = getOption("dwtools.db.auto.table.name.nchar"))
}
\arguments{
\item{x}{character names of data.table.}

\item{ncol}{integer number of cols to use.}

\item{nchar}{integer number of characters of each col to use.}
}
\description{
Automatic table name generation based on first ncol and first nchar of those cols, all pasted, suffix added as timestamp*1e3.
}

