% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CJI}
\alias{CJI}
\title{Indexed cross join / filter for data.table}
\usage{
CJI(IDX, ..., nomatch = 0)
}
\arguments{
\item{IDX}{list generated by \link{idxv}, must be produced on the same data.table on which \emph{CJI} is used.}

\item{nomatch}{integer \emph{0} or \emph{NA}. Default \emph{0} (inner join) to filter data.table.}

\item{\dots}{values to filter data.table. For the columns to skip filtering use \code{TRUE}, see examples in \link{idxv}.}
}
\value{
Subset of data.table according to arguments passed to \dots.
}
\description{
To be used inside data.table, same as \link{CJ} function. In case of not filtering on particular fields pass \code{TRUE} value to \dots, see examples in \link{idxv}.
}
\examples{
# see ?idxv examples
}
\seealso{
\link{idxv}
}

