% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data.equal.data.table}
\alias{data.equal.data.table}
\title{Data equal in two data.tables}
\usage{
data.equal.data.table(DT1, DT2, ignore_row_order = TRUE,
  ignore_col_order = FALSE, check.attributes = FALSE)
}
\arguments{
\item{DT1}{data.table.}

\item{DT2}{data.table.}

\item{ignore_row_order}{logical.}

\item{ignore_col_order}{logical.}

\item{check.attributes}{logical, only \emph{FALSE} supported.}
}
\description{
Test if data equal in two data.tables, can ignore order of rows or columns.
}
\note{
Duplicate names in DTs were not tested. All attributes all ignored.
}
\examples{
suppressPackageStartupMessages(library(data.table))
library(dwtools)

DT1 <- dw.populate(scenario="fact")
DT2 <- copy(DT1)[sample(1:nrow(DT1),nrow(DT1),FALSE)]
data.equal.data.table(DT1,DT2,ignore_row_order=TRUE)

setcolorder(DT2,sample(names(DT2),length(DT2),FALSE))
data.equal.data.table(DT1,DT2,ignore_row_order=TRUE,ignore_col_order=TRUE)
}

