% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db_dict.R
\name{db_dict}
\alias{db_dict}
\title{db_dict}
\usage{
db_dict()
}
\description{
Dictionary for databases drivers write, read, get, send, schematable, preprocess, postprocess. Extend here for new db/interface (PR welcome). Live extend by option \code{options("dwtools.db.dict"=db_dict())}.
}
\keyword{internal}

