% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbCopy}
\alias{dbCopy}
\title{Copy tables between databases}
\usage{
dbCopy(source.table.name, source.conn.name, target.table.name, target.conn.name,
  verbose = getOption("dwtools.verbose"))
}
\arguments{
\item{source.table.name}{}
}
\description{
Copy tables between databases
}

