% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtt.R
\name{dtt}
\alias{dtt}
\title{data.table timing}
\usage{
dtt(unchain = TRUE, arg.names = TRUE, purge = FALSE)
}
\arguments{
\item{unchain}{logical print only single chain elements, default \emph{TRUE}. When \emph{FALSE} it will print direct call.}

\item{arg.names}{logical, print also \code{DT[i,j,...]} arguments, affects only \code{unchain=TRUE}.}

\item{purge}{logical truncate current data.table timing logs.}
}
\description{
data.table timing
}
\examples{
suppressPackageStartupMessages(library(dwtools)); dtt(purge=TRUE); options("datatable.timing"=FALSE)

DT <- data.table(a = 1L:5L)
DT[1:4,list(a)][2:3,list(a)]
dtt() # no timing yet

options("datatable.timing"=TRUE)
DT[1:4,list(a)][2:3,list(a)]
dtt()
dtt(arg.names=FALSE)
dtt(unchain=FALSE)

# measure overhead
dtt(purge=TRUE)
N <- 1e3
options("datatable.timing"=FALSE)
system.time(for(i in 1:N) DT[1:4,list(a)][2:3,list(a)])

options("datatable.timing"=TRUE)
system.time(for(i in 1:N) DT[1:4,list(a)][2:3,list(a)])

dtt()
dtt(purge=TRUE)
dtt()
}

