% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dw.explore}
\alias{dw.explore}
\title{DW explore}
\usage{
dw.explore(x, filter = list(quantile = 0.025, nrow.ratio = 0.05,
  hierarchical.redundancy = TRUE), materialize = FALSE,
  materialize.FUN = sum,
  .db.conn.name = names(getOption("dwtools.db.conns"))[1],
  timing = getOption("dwtools.timing"))
}
\arguments{
\item{x}{data.table}

\item{filter}{list}

\item{materialize}{logical}

\item{materialize.FUN}{function}

\item{.db.conn.name}{character}

\item{timing}{logical}
}
\description{
Explore hierarchies in data.table by groupings combinations and cardinality check. To improve performance the function uses internal data.table 1.9.6 functions.
}
\examples{
suppressPackageStartupMessages(library(data.table))
library(dwtools)

X = dw.populate(scenario="star schema")
x <- joinbyv(X$SALES, join=list(X$CURRENCY, X$GEOGRAPHY))

r <- timing(dw.explore(x, filter=list("quantile"=0.025,"nrow.ratio"=0.05,"hierarchical.redundancy"=TRUE)),
            NA_integer_,"dw.explore")
attr(r,"timing")

if(interactive()) shiny::runApp("dwtools/inst/shinyDW")
}
\seealso{
\link{joinbyv}
}

