% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dw.populate}
\alias{dw.populate}
\title{Populate DW data}
\usage{
dw.populate(N = 1e+05, K = 100, S = 1, scenario = "star",
  setkey = TRUE, verbose = getOption("dwtools.verbose"))
}
\arguments{
\item{N}{integer facts volume}

\item{K}{groups size}

\item{S}{integer used in \link{set.seed}}

\item{scenario}{character in \code{c("fact","star","denormalize")}}

\item{setkey}{logical used only for scenario \code{"star"}, default TRUE makes dimensions populated with keys already. For non-key benchmarks use FALSE, also remember about \code{getOption("datatable.auto.index"=FALSE)},}

\item{verbose}{integer print sub statuses}
}
\description{
Populate sample DW data. By default \emph{star}, see \code{scenario} arg for others.
}
\details{
The following case \code{scenario="denormalize"} will invoke lookup for full columns set in all the dimensions. On the real data it is advised to use \code{scenario="star"} and later denormalize using \link{joinbyv} function where you can provide subsets of columns on each lookup.
}
\examples{
suppressPackageStartupMessages(library(dwtools))

X = dw.populate() # scenario="star"
lapply(X, head)

Z = dw.populate(scenario="denormalize")
print(Z)

SALES = dw.populate(scenario="fact")
print(SALES)
}
\seealso{
\link{joinbyv}
}

