% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dwtools.R
\docType{package}
\name{dwtools}
\alias{dwtools}
\alias{dwtools-package}
\title{Data Warehouse tools}
\description{
Extension for \link{data.table} package for Data Warehouse related functionalities.
}
\details{
The core functions includes:
\itemize{
\item \link{db} as Extracting and Loading tool in ETL terms.
\item \link{CUBE} cube class object and MDX like queries on it.
\item \link{joinbyv} a denormalization of star schema and snowflake schema to single table.
\item \link{dw.populate} populate star schema data.
\item \link{timing} measure timing and rows in-out, including logging to db and verbose messages.
\item \link{build_hierarchy} transform single dataset to star schema, columns allocation based on cardinalities in unique groupings of each pair.
\item \link{idxv} custom indices for in-memory processing.
}
}
\note{
All dot prefixed arguments are designed to be taken from the options, use them only in special cases, they may be removed from functions input args in future.\cr
The base option \emph{digits.secs} will be changed to 3 on package load.
}

