% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timing.R
\name{get.timing}
\alias{get.timing}
\title{Get timing logs}
\usage{
get.timing(trunc_expr = 16L, last = FALSE)
}
\arguments{
\item{trunc_expr}{integer or logical, when \emph{integer} then cut \emph{expr} field after nchar, or when \emph{TRUE} remove \emph{expr} from result, when \emph{FALSE} include full \emph{expr}.}

\item{last}{integer or logical, if \emph{FALSE} then full log returned, when \emph{TRUE} then only last log entry is returned, when integer then last n entries returned.}
}
\description{
Get timing logs
}
\seealso{
\link{timing}, \link{purge.timing}
}

