% Generated by roxygen2 (4.0.2): do not edit by hand
\name{int.is.POSIXct}
\alias{int.is.POSIXct}
\title{int.is.POSIXct}
\usage{
int.is.POSIXct(x, date_from = as.POSIXct("1970-01-01"),
  date_to = as.POSIXct("2100-01-01"))
}
\description{
Check if is integer and can be POSIX between 1970 and 2100.
}
\keyword{internal}

