% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.int.POSIXct}
\alias{is.int.POSIXct}
\title{is.int.POSIXct}
\usage{
is.int.POSIXct(x, date_from = as.POSIXct("1970-01-01"),
  date_to = as.POSIXct("2100-01-01"))
}
\description{
Check if is integer and could be POSIX between 1970 and 2100.
}
\keyword{internal}

