% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{is.table.name}
\alias{is.table.name}
\title{is.table.name}
\usage{
is.table.name(x)
}
\description{
test if character is.table.name (no spaces), accept vector, return scalar logical.
}
\keyword{internal}

