% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.unique}
\alias{is.unique}
\title{Fast check is vector unique}
\usage{
is.unique(x)
}
\arguments{
\item{x}{vector}
}
\description{
Test if vector unique.
}
\examples{
suppressPackageStartupMessages(library(dwtools))

x1 <- c("a","b","c","d","e")
x2 <- c(x1,"d")
is.unique(x1)
is.unique(x2)

# my previous approach
is.unq <- function(x) length(x)==length(unique(x))

x <- sample(1e2, 1e7, TRUE) # non-unq
system.time(is.unq(x))
system.time(is.unique(x))
x <- sample(1e7, 1e7, FALSE) # unq
system.time(is.unq(x))
system.time(is.unique(x))
}

