% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{list.sub}
\alias{list.sub}
\title{list.sub}
\usage{
list.sub(x, i, fill = NULL)
}
\description{
Take \code{i} element of the list \code{x} if not present or NULL then return \code{fill}. Used in \code{db} for \dots argument processing.
}
\keyword{internal}

