% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgsVersion}
\alias{pkgsVersion}
\title{pkgsVersion}
\usage{
pkgsVersion(pkgs, libs = .libPaths())
}
\arguments{
\item{pkgs}{character vector of packages names.}

\item{libs}{character vector of libraries paths to compare, vector names will be column names.}
}
\description{
Batch package version compare between libraries.
}
\examples{
suppressPackageStartupMessages(library(dwtools))

pkgs <- c("dplyr","data.table","RSQLite","dwtools")
pkgsVersion(pkgs)

# custom libraries
pkgsVersion(pkgs, libs = c(dev = .libPaths()[1], prod = .libPaths()[2]))
}

