% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vwap.R
\name{round.POSIXct}
\alias{round.POSIXct}
\title{round.POSIXct}
\usage{
\method{round}{POSIXct}(x, units, trunc_ceiling = "trunc", tz = "UTC")
}
\description{
Trunc or ceiling of POSIXct to particular time units
}
\keyword{internal}

