% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.table.xts}
\alias{as.data.table.xts}
\title{as.data.table.xts}
\usage{
\method{as.data.table}{xts}(x, keep.rownames = TRUE)
}
\arguments{
\item{x}{xts to convert to data.table}

\item{keep.rownames}{keep xts index as \emph{index} column in result data.table}
}
\description{
convert xts to data.table
}
\examples{
suppressPackageStartupMessages(library(data.table))
library(dwtools)
options("dwtools.verbose"=3)  # turn on status messages printed to console

data(sample_matrix, package = "xts")
sample.xts <- xts::as.xts(sample_matrix)
# print head of xts
print(head(sample.xts))
# print dt
print(as.data.table(sample.xts))
}
\seealso{
\link{as.xts.data.table}
}

