% Generated by roxygen2 (4.0.2): do not edit by hand
\name{auto.table.name}
\alias{auto.table.name}
\title{auto.table.name}
\usage{
auto.table.name(x, ncol = getOption("dwtools.db.auto.table.name.ncol"),
  nchar = getOption("dwtools.db.auto.table.name.nchar"))
}
\description{
Automatic table name generation based on first ncols and first nchars of those cols, all pasted, suffix added as timestamp*1e3.
}
\keyword{internal}

