% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dbCopy}
\alias{dbCopy}
\title{Copy tables between databases}
\usage{
dbCopy(source.table.name, source.conn.name, target.table.name, target.conn.name,
  timing = getOption("dwtools.timing"),
  verbose = getOption("dwtools.verbose"))
}
\arguments{
\item{source.table.name}{character vector of tables names to copy from source connection}

\item{source.conn.name}{character scalar}

\item{target.table.name}{character vector of tables names to copy to target connection}

\item{target.conn.name}{character scalar}

\item{timing}{logical measure timing}

\item{verbose}{integer status messages}
}
\description{
Copy tables between databases
}
\examples{
# see the last example in ?db
}
\seealso{
\link{db}, \link{timing}
}

