% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dw.populate}
\alias{dw.populate}
\title{Populate data}
\usage{
dw.populate(n = 1000, seed = 1, scenario = c("fact", "star schema",
  "star schema denormalized"), verbose = getOption("dwtools.verbose"))
}
\arguments{
\item{n}{integer facts volume}

\item{seed}{integer used in \code{set.seed}}

\item{scenario}{character in \code{c("fact","star schema","star schema denormalized")}}

\item{verbose}{integer print sub statuses}
}
\description{
sample DW data. TODO better example data, k argument for groups size like in DT benchmark
}

