% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{dwtools}
\alias{dwtools}
\alias{dwtools-package}
\title{dwtools}
\description{
Extension for \link{data.table} package for Data Warehouse related functionalities.
}
\references{
\url{https://www.github.com/jangorecki/dwtools}
}

