% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timing}
\alias{timing}
\title{Measure timing}
\usage{
timing(expr, nrow_in = NA_integer_, .timing = getOption("dwtools.timing"),
  .timing.name = getOption("dwtools.timing.name"),
  .timing.conn.name = getOption("dwtools.timing.conn.name"))
}
\arguments{
\item{expr}{expression}

\item{nrow_in}{integer manually provided input object nrow}

\item{.timing}{logical}

\item{.timing.name}{character}

\item{.timing.conn.name}{character}
}
\description{
Collect timings and nrows when possible.
}
\details{
Use option \code{options("dwtools.timing"=TRUE)} to turn on timing measurment in functions which supports timing measurement. To log timing to db connection, setup \code{options("dwtools.db.conns"}, provide connectio name to \code{options("dwtools.timing.conn.name"="sqlite1")} and target table \code{options("dwtools.timing.name"="mylogtable")} otherwise timing will be returned as \code{"timing"} attribute to the expression result.
}

